Support create java function project (#66)
This leverages [maven archetype](https://github.com/Microsoft/azure-maven-archetypes/tree/master/azure-functions-archetype).

+++ src/commands/createNewProject.ts
@@ -8,7 +8,10 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import { OutputChannel } from 'vscode';
 import { IUserInterface } from '../IUserInterface';
+import { Pick } from '../IUserInterface';
 import { localize } from '../localize';
+import { TemplateLanguage } from '../templates/Template';
+import { cpUtils } from '../utils/cpUtils';
 import { confirmOverwriteFile } from '../utils/fs';
 import * as fsUtil from '../utils/fs';
 import { gitUtils } from '../utils/gitUtils';
@@ -16,7 +19,25 @@ import * as workspaceUtil from '../utils/workspace';
 import { VSCodeUI } from '../VSCodeUI';
 
 const taskId: string = 'launchFunctionApp';
-const tasksJson: {} = {
+
+const problemMatcher: {} = {
+    owner: 'azureFunctions',
+    pattern: [
+        {
+            regexp: '\\b\\B',
+            file: 1,
+            location: 2,
+            message: 3
+        }
+    ],
+    background: {
+        activeOnStart: true,
+        beginsPattern: '^.*Stopping host.*',
+        endsPattern: '^.*Job host started.*'
+    }
+};
+
+const tasksJsonForJavaScript: {} = {
     version: '2.0.0',
     tasks: [
         {
@@ -29,28 +50,40 @@ const tasksJson: {} = {
                 reveal: 'always'
             },
             problemMatcher: [
-                {
-                    owner: 'azureFunctions',
-                    pattern: [
-                        {
-                            regexp: '\\b\\B',
-                            file: 1,
-                            location: 2,
-                            message: 3
-                        }
-                    ],
-                    background: {
-                        activeOnStart: true,
-                        beginsPattern: '^.*Stopping host.*',
-                        endsPattern: '^.*Job host started.*'
-                    }
-                }
+                problemMatcher
             ]
         }
     ]
 };
 
-const launchJson: {} = {
+const tasksJsonForJava: {} = {
+    version: '2.0.0',
+    tasks: [
+        {
+            label: localize('azFunc.launchFuncApp', 'Launch Function App'),
+            identifier: taskId,
+            linux: {
+                command: 'sh -c "mvn clean package -B && func host start --script-root %path%"'
+            },
+            osx: {
+                command: 'sh -c "mvn clean package -B && func host start --script-root %path%"'
+            },
+            windows: {
+                command: 'powershell mvn clean package -B; func host start --script-root %path%'
+            },
+            type: 'shell',
+            isBackground: true,
+            presentation: {
+                reveal: 'always'
+            },
+            problemMatcher: [
+                problemMatcher
+            ]
+        }
+    ]
+};
+
+const launchJsonForJavaScript: {} = {
     version: '0.2.0',
     configurations: [
         {
@@ -59,6 +92,20 @@ const launchJson: {} = {
             request: 'attach',
             port: 5858,
             protocol: 'inspector',
+            preLaunchTask: taskId
+        }
+    ]
+};
+
+const launchJsonForJava: {} = {
+    version: '0.2.0',
+    configurations: [
+        {
+            name: localize('azFunc.attachToFunc', 'Attach to Azure Functions'),
+            type: 'java',
+            request: 'attach',
+            hostName: 'localhost',
+            port: 5005,
             preLaunchTask: taskId
         }
     ]
@@ -100,9 +147,83 @@ const localSettingsJson: {} = {
     }
 };
 
+async function promotForMavenParameters(ui: IUserInterface, functionAppPath: string): Promise<IMavenParameters> {
+    const groupIdPlaceHolder: string = localize('azFunc.java.groupIdPlaceholder', 'Group ID');
+    const groupIdPrompt: string = localize('azFunc.java.groupIdPrompt', 'Provide value for groupId');
+    const groupId: string = await ui.showInputBox(groupIdPlaceHolder, groupIdPrompt, false, undefined, 'com.function');
+
+    const artifactIdPlaceHolder: string = localize('azFunc.java.artifactIdPlaceholder', 'Artifact ID');
+    const artifactIdprompt: string = localize('azFunc.java.artifactIdPrompt', 'Provide value for artifactId');
+    const artifactId: string = await ui.showInputBox(artifactIdPlaceHolder, artifactIdprompt, false, undefined, path.basename(functionAppPath));
+
+    const versionPlaceHolder: string = localize('azFunc.java.versionPlaceHolder', 'Version');
+    const versionPrompt: string = localize('azFunc.java.versionPrompt', 'Provide value for version');
+    const version: string = await ui.showInputBox(versionPlaceHolder, versionPrompt, false, undefined, '1.0-SNAPSHOT');
+
+    const packagePlaceHolder: string = localize('azFunc.java.packagePlaceHolder', 'Package');
+    const packagePrompt: string = localize('azFunc.java.packagePrompt', 'Provide value for package');
+    const packageName: string = await ui.showInputBox(packagePlaceHolder, packagePrompt, false, undefined, groupId);
+
+    const appNamePlaceHolder: string = localize('azFunc.java.appNamePlaceHolder', 'App Name');
+    const appNamePrompt: string = localize('azFunc.java.appNamePrompt', 'Provide value for appName');
+    const appName: string = await ui.showInputBox(appNamePlaceHolder, appNamePrompt, false, undefined, `${artifactId}-${Date.now()}`);
+
+    return {
+        groupId: groupId,
+        artifactId: artifactId,
+        version: version,
+        packageName: packageName,
+        appName: appName
+    };
+}
+
 export async function createNewProject(outputChannel: OutputChannel, functionAppPath?: string, openFolder: boolean = true, ui: IUserInterface = new VSCodeUI()): Promise<void> {
     if (functionAppPath === undefined) {
         functionAppPath = await workspaceUtil.selectWorkspaceFolder(ui, localize('azFunc.selectFunctionAppFolderNew', 'Select the folder that will contain your function app'));
+    }
+
+    const languages: Pick[] = [
+        new Pick(TemplateLanguage.JavaScript),
+        new Pick(TemplateLanguage.Java)
+    ];
+    const language: string = (await ui.showQuickPick(languages, localize('azFunc.selectFuncTemplate', 'Select a language for your function project'))).label;
+
+    let javaTargetPath: string = '';
+    switch (language) {
+        case TemplateLanguage.Java:
+            // Get parameters for Maven command
+            const { groupId, artifactId, version, packageName, appName } = await promotForMavenParameters(ui, functionAppPath);
+            // Use maven command to init Java function project.
+            await cpUtils.executeCommand(
+                outputChannel,
+                functionAppPath,
+                'mvn',
+                'archetype:generate',
+                '-DarchetypeGroupId="com.microsoft.azure"',
+                '-DarchetypeArtifactId="azure-functions-archetype"',
+                `-DgroupId="${groupId}"`,
+                `-DartifactId="${artifactId}"`,
+                `-Dversion="${version}"`,
+                `-Dpackage="${packageName}"`,
+                `-DappName="${appName}"`,
+                '-B' // in Batch Mode
+            );
+
+            functionAppPath = path.join(functionAppPath, artifactId);
+            javaTargetPath = `target/azure-functions/${appName}/`;
+            break;
+        default:
+            // the maven archetype contains these files, so not check them when language is Java
+            const hostJsonPath: string = path.join(functionAppPath, 'host.json');
+            if (await confirmOverwriteFile(hostJsonPath)) {
+                await fsUtil.writeFormattedJson(hostJsonPath, hostJson);
+            }
+
+            const localSettingsJsonPath: string = path.join(functionAppPath, 'local.settings.json');
+            if (await confirmOverwriteFile(localSettingsJsonPath)) {
+                await fsUtil.writeFormattedJson(localSettingsJsonPath, localSettingsJson);
+            }
+            break;
     }
 
     const vscodePath: string = path.join(functionAppPath, '.vscode');
@@ -112,33 +233,51 @@ export async function createNewProject(outputChannel: OutputChannel, functionApp
         await gitUtils.gitInit(outputChannel, functionAppPath);
 
         const gitignorePath: string = path.join(functionAppPath, '.gitignore');
-        if (await confirmOverwriteFile(gitignorePath)) {
+        if (language !== TemplateLanguage.Java && await confirmOverwriteFile(gitignorePath)) {
             await fse.writeFile(gitignorePath, gitignore);
         }
     }
 
     const tasksJsonPath: string = path.join(vscodePath, 'tasks.json');
     if (await confirmOverwriteFile(tasksJsonPath)) {
-        await fsUtil.writeFormattedJson(tasksJsonPath, tasksJson);
+        switch (language) {
+            case TemplateLanguage.Java:
+                let tasksJsonString: string = JSON.stringify(tasksJsonForJava);
+                tasksJsonString = tasksJsonString.replace(/%path%/g, javaTargetPath);
+                // tslint:disable-next-line:no-string-literal no-unsafe-any
+                const tasksJson: {} = JSON.parse(tasksJsonString);
+                // tslint:disable-next-line:no-string-literal no-unsafe-any
+                tasksJson['tasks'][0]['problemMatcher'][0]['background']['beginsPattern'] = '^.*Scanning for projects.*';
+                await fsUtil.writeFormattedJson(tasksJsonPath, tasksJson);
+                break;
+            default:
+                await fsUtil.writeFormattedJson(tasksJsonPath, tasksJsonForJavaScript);
+                break;
+        }
     }
 
     const launchJsonPath: string = path.join(vscodePath, 'launch.json');
     if (await confirmOverwriteFile(launchJsonPath)) {
-        await fsUtil.writeFormattedJson(launchJsonPath, launchJson);
-    }
-
-    const hostJsonPath: string = path.join(functionAppPath, 'host.json');
-    if (await confirmOverwriteFile(hostJsonPath)) {
-        await fsUtil.writeFormattedJson(hostJsonPath, hostJson);
-    }
-
-    const localSettingsJsonPath: string = path.join(functionAppPath, 'local.settings.json');
-    if (await confirmOverwriteFile(localSettingsJsonPath)) {
-        await fsUtil.writeFormattedJson(localSettingsJsonPath, localSettingsJson);
+        switch (language) {
+            case TemplateLanguage.Java:
+                await fsUtil.writeFormattedJson(launchJsonPath, launchJsonForJava);
+                break;
+            default:
+                await fsUtil.writeFormattedJson(launchJsonPath, launchJsonForJavaScript);
+                break;
+        }
     }
 
     if (openFolder && !workspaceUtil.isFolderOpenInWorkspace(functionAppPath)) {
         // If the selected folder is not open in a workspace, open it now. NOTE: This may restart the extension host
         await vscode.commands.executeCommand('vscode.openFolder', vscode.Uri.file(functionAppPath), false);
     }
+}
+
+interface IMavenParameters {
+    groupId: string;
+    artifactId: string;
+    version: string;
+    packageName: string;
+    appName: string;
 }
